+++ map/index.cjs
@@ -1,2 +1,0 @@
-exports = module.exports = require('../commonjs/read/schema/mapToObjects.js').default
-exports['default'] = require('../commonjs/read/schema/mapToObjects.js').default