+++ node/index.d.ts
@@ -6,7 +6,6 @@ import { Stream } from 'stream';
 
 import {
 	ParseWithSchemaOptions,
-	ParseWithMapOptions,
 	ParseWithoutSchemaOptions,
 	ParsedObjectsResult,
 	Row
@@ -28,7 +27,6 @@ export function parseExcelDate(excelSerialDate: number) : typeof Date;
 type Input = Stream | Buffer | PathLike;
 
 export function readXlsxFile<T extends object>(input: Input, options: ParseWithSchemaOptions<T>) : Promise<ParsedObjectsResult<T>>;
-export function readXlsxFile<T extends object>(input: Input, options: ParseWithMapOptions) : Promise<ParsedObjectsResult<T>>;
 export function readXlsxFile(input: Input, options?: ParseWithoutSchemaOptions) : Promise<Row[]>;
 
 export function readSheetNames(input: Input) : Promise<string[]>;