+++ web-worker/index.d.ts
@@ -1,6 +1,5 @@
 import {
 	ParseWithSchemaOptions,
-	ParseWithMapOptions,
 	ParseWithoutSchemaOptions,
 	ParsedObjectsResult,
 	Row
@@ -22,7 +21,6 @@ export function parseExcelDate(excelSerialDate: number) : typeof Date;
 type Input = File | Blob | ArrayBuffer;
 
 export function readXlsxFile<T extends object>(input: Input, options: ParseWithSchemaOptions<T>) : Promise<ParsedObjectsResult<T>>;
-export function readXlsxFile<T extends object>(input: Input, options: ParseWithMapOptions) : Promise<ParsedObjectsResult<T>>;
 export function readXlsxFile(input: Input, options?: ParseWithoutSchemaOptions) : Promise<Row[]>;
 
 export function readSheetNames(input: Input) : Promise<string[]>;